+++ src/main/java/com/sogeti/SoSenseiPsyche/model/CodeGenerator.java
@@ -1,0 +1,4 @@
+package com.sogeti.SoSenseiPsyche.model;
+
+public class CodeGenerator {
+}
