+++ src/main/java/com/sogeti/SoSenseiPsyche/model/Feedback.java
@@ -1,0 +1,4 @@
+package com.sogeti.SoSenseiPsyche.model;
+
+public class Feedback {
+}
